Added support for all pivot functions when calculating a pivot table

+++ src/EPPlus/FormulaParsing/Excel/Functions/MathFunctions/Stdev.cs
@@ -61,7 +61,7 @@ namespace OfficeOpenXml.FormulaParsing.Excel.Functions.MathFunctions
                 {
                     return CompileResult.GetErrorResult(eErrorType.Div0);
                 }
-                ret = MathObj.Sqrt((double)div);
+                ret = MathObj.Sqrt(div);
             }
             return CreateResult(ret, DataType.Decimal);
         }
